krishi_sakha_py/configs/supabase_key.py
import supabase
import os
from dotenv import load_dotenv

load_dotenv()

SUPABASE_URL = os.getenv("URL")
print(SUPABASE_URL)

SUPABASE_SERVICE_ROLE_KEY = os.getenv("SERVICE_ROLE_KEY")
print(SUPABASE_SERVICE_ROLE_KEY)
SUPABASE_ANON_KEY = os.getenv("ANON_PUBLIC_KEY")
print(SUPABASE_ANON_KEY)
SUPABASE_SECRET_KEY = os.getenv("SECRET_KEY")
print(SUPABASE_SECRET_KEY)
SUPABASE_LEGACY_JWT_KEY = os.getenv("LEGACY_JWT_KEY")
print(SUPABASE_LEGACY_JWT_KEY)
SUPABASE = supabase.create_client(SUPABASE_URL, SUPABASE_SERVICE_ROLE_KEY)


